pyproject.toml
[tool.poetry]
name = "aucampia.cookiecutter-python"
version = "0.0.0"
description = ""
license = "CC0-1.0"
readme = "README.md"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
codespell = "^2.2.6"
# flake8 = "^6.1.0"
# # flake8-black = ">=0"
# flake8-bugbear = "^23.9.16"
# flake8-isort = "^6.1.0"
# flake8-bandit = "^4.1.1"
# isort = "^5.12.0"
mypy = "^1.6.1"
# pep8-naming = "^0.13.3"
# pycln = "^2.3.0"
pytest = "^7.4.2"
pytest-cookies = "^0.7.0"
pytest-cov = "^4.1.0"
types-PyYAML = "^6.0.12.12"
poetry = "^1.6.1"
cookiecutter = "^2.4.0"
frozendict = "^2.3.8"
types-frozendict = "^2.0.9"
pip-audit = "^2.6.1"
ruff = "^0.1.0"


[tool.poetry.group.templates.dependencies]
typer = "^0.9.0"
structlog = "^23.2.0"

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
preview = true
target-version = "py39"
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "N", # pep8-naming
    "RUF", # Ruff
    "UP", # pyupgrade
    "FA", # flake8-future-annotations
]

ignore = [
    "E501", # line too long:
    # Disabled based on black recommendations
    # https://black.readthedocs.io/en/stable/faq.html#why-are-flake8-s-e203-and-w503-violated
    "E203", # whitespace before ':'
    "E231", # missing whitespace after ','
    "E251", # Unexpected spaces around keyword / parameter equals
]

extend-exclude = [
    "examples/**",
]

# Same as Black.
line-length = 88

[tool.ruff.per-file-ignores]


[tool.coverage.report]
# https://coverage.readthedocs.io/en/coverage-5.0/config.html
show_missing = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/customize.html
# https://docs.pytest.org/en/stable/reference.html#configuration-options
testpaths=[
    "tests"
]
addopts = [
    "--cov-config=pyproject.toml",
    "--capture=no",
    "--tb=native",
    "-ra",
]
log_format = "%(asctime)s %(process)d %(thread)d %(levelno)03d:%(levelname)-8s %(name)-12s %(module)s:%(lineno)s:%(funcName)s %(message)s"
log_date_format = "%Y-%m-%dT%H:%M:%S"
log_cli_format = "%(asctime)s %(process)d %(thread)d %(levelno)03d:%(levelname)-8s %(name)-12s %(module)s:%(lineno)s:%(funcName)s %(message)s"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S"

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
files = "hooks,tests,link_project/pkg_files"
python_version = "3.10"
strict = true
warn_unreachable = true
warn_unused_configs = true
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest_cookies.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cookiecutter.*"
ignore_missing_imports = true

[tool.pycln]
# https://hadialqattan.github.io/pycln/#/?id=cli-options
# https://hadialqattan.github.io/pycln/#/?id=example
paths = ["hooks", "tests", "link_project/pkg_files"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

